Refactor resume management into sub component of AppSidebar
