+++ src/components/AppSidebar.tsx
@@ -1,26 +1,14 @@
-import React, { useState, useRef, useEffect } from "react";
+import React from "react";
 import { Sidebar, SidebarContent, SidebarFooter, SidebarGroup, SidebarGroupContent, SidebarGroupLabel, SidebarHeader, SidebarMenu, SidebarMenuButton, SidebarMenuItem, SidebarTrigger } from "@/components/ui/sidebar"
 import SortableNav from "./SortableNav";
-import { DownloadCloud, EraserIcon, ExternalLink, FileJson, FlaskConical, ListPlus, UploadCloud, Edit2, Check, X, MoreVertical, Copy, FileText, Trash2, SaveAllIcon, FolderOpen, Sheet, File } from "lucide-react";
+import { DownloadCloud, EraserIcon, ExternalLink, FileJson, FlaskConical, ListPlus, UploadCloud } from "lucide-react";
 import Corgi from "./Corgi";
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from "./ui/dialog";
 import { ResumeImporter } from "./ResumeImporter";
 import { FormData } from "@/types";
 import { TemplateSwitcher } from "./TemplateSwitcher";
 import { useResume } from '@/lib/ResumeContext';
-import Button from "./Button";
-import { 
-  DropdownMenu, 
-  DropdownMenuContent, 
-  DropdownMenuItem, 
-  DropdownMenuLabel, 
-  DropdownMenuSeparator, 
-  DropdownMenuTrigger 
-} from "./ui/dropdown-menu";
-import { Checkbox } from "./ui/checkbox";
-import { getSavedResumes, saveResumeCopy, loadResumeCopy, deleteResumeCopy } from '@/lib/StorageService';
-import { createSectionsFromFormData } from '@/lib/DataInitializer';
-import { toast } from "sonner";
+import ResumeManager from "./ResumeManager";
 
 interface SidebarProps {
   resetData?: () => void;
@@ -41,192 +29,17 @@ const exportJson = (e: React.MouseEvent<HTMLAnchorElement, MouseEvent>, onExport
 
 const corgiSize: number = 84;
 
-// Format relative time like "3 days ago", "2 weeks ago", etc.
-const formatRelativeTime = (dateString: string): string => {
-  const date = new Date(dateString);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-  
-  if (diffDays === 0) {
-    const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-    if (diffHours === 0) {
-      const diffMinutes = Math.floor(diffMs / (1000 * 60));
-      if (diffMinutes === 0) return 'just now';
-      if (diffMinutes === 1) return '1 minute ago';
-      return `${diffMinutes} minutes ago`;
-    }
-    if (diffHours === 1) return '1 hour ago';
-    return `${diffHours} hours ago`;
-  }
-  
-  if (diffDays === 1) return 'yesterday';
-  if (diffDays < 7) return `${diffDays} days ago`;
-  
-  const diffWeeks = Math.floor(diffDays / 7);
-  if (diffWeeks === 1) return '1 week ago';
-  if (diffWeeks < 4) return `${diffWeeks} weeks ago`;
-  
-  // More than a month, show the actual date
-  return date.toLocaleDateString();
-};
-
 function AppSidebar({
   resetData,
   sampleData,
   onExport,
   onImportJsonFormData,
 }: SidebarProps) {
-  const { formData, sections, addGenericSection, resumeName, setResumeName, setFormData, setSections, selectedTemplate } = useResume();
-  const [isEditingName, setIsEditingName] = useState(false);
-  const [tempName, setTempName] = useState(resumeName);
-  const [savedResumes, setSavedResumes] = useState(getSavedResumes());
-  const [isOpenDialogOpen, setIsOpenDialogOpen] = useState(false);
-  const [selectedResumeIds, setSelectedResumeIds] = useState<Set<string>>(new Set());
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    setTempName(resumeName);
-  }, [resumeName]);
-
-  useEffect(() => {
-    if (isEditingName && inputRef.current) {
-      inputRef.current.focus();
-      inputRef.current.select();
-    }
-  }, [isEditingName]);
-
-  const handleStartEdit = () => {
-    setTempName(resumeName);
-    setIsEditingName(true);
-  };
-
-  const handleSaveEdit = () => {
-    const trimmedName = tempName.trim();
-    if (trimmedName && trimmedName !== resumeName) {
-      setResumeName(trimmedName);
-    }
-    setIsEditingName(false);
-  };
-
-  const handleCancelEdit = () => {
-    setTempName(resumeName);
-    setIsEditingName(false);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === 'Enter') {
-      e.preventDefault();
-      handleSaveEdit();
-    } else if (e.key === 'Escape') {
-      e.preventDefault();
-      handleCancelEdit();
-    }
-  };
-
-  const handleSaveAsCopy = () => {
-    // Generate auto-incrementing name like "resume-name (1)", "resume-name (2)", etc.
-    const baseName = resumeName.replace(/ \(\d+\)$/, ''); // Remove existing (n) suffix if present
-    const existingResumes = getSavedResumes();
-    const existingNames = existingResumes.map(r => r.name);
-    
-    let copyNumber = 1;
-    let newName = `${baseName} (${copyNumber})`;
-    
-    // Find the next available number
-    while (existingNames.includes(newName)) {
-      copyNumber++;
-      newName = `${baseName} (${copyNumber})`;
-    }
-    
-    // Save the copy
-    const resumeId = saveResumeCopy({ formData, sections, templateId: selectedTemplate.id }, newName);
-    if (resumeId) {
-      // Immediately switch to the new copy
-      setResumeName(newName);
-      setSavedResumes(getSavedResumes());
-      toast.success(`Saved as copy: ${newName}`);
-    }
-  };
-
-  const handleLoadResume = (resumeId: string) => {
-    if (window.confirm('Loading this resume will replace your current work. Continue?')) {
-      const resumeData = loadResumeCopy(resumeId);
-      if (resumeData) {
-        setFormData(resumeData.formData);
-        setSections(resumeData.sections);
-        // Update the name from the saved copy
-        const savedCopy = savedResumes.find(r => r.id === resumeId);
-        if (savedCopy) {
-          setResumeName(savedCopy.name);
-        }
-        setIsOpenDialogOpen(false);
-      }
-    }
-  };
+  const { addGenericSection } = useResume();
 
   const handleNewResume = () => {
     if (window.confirm('Creating a new resume will clear your current work. Continue?')) {
       resetData?.();
-    }
-  };
-
-  const handleDeleteResume = (resumeId: string, resumeName: string) => {
-    if (window.confirm(`Are you sure you want to delete "${resumeName}"?`)) {
-      if (deleteResumeCopy(resumeId)) {
-        setSavedResumes(getSavedResumes());
-      }
-    }
-  };
-
-  // Refresh saved resumes when dialog opens and clear selections
-  useEffect(() => {
-    if (isOpenDialogOpen) {
-      setSavedResumes(getSavedResumes());
-      setSelectedResumeIds(new Set());
-    }
-  }, [isOpenDialogOpen]);
-
-  const handleDeleteSelected = () => {
-    const selectedCount = selectedResumeIds.size;
-    if (selectedCount === 0) return;
-    
-    const message = selectedCount === 1 
-      ? 'Are you sure you want to delete this resume?'
-      : `Are you sure you want to delete ${selectedCount} resumes?`;
-    
-    if (window.confirm(message)) {
-      let deletedCount = 0;
-      selectedResumeIds.forEach(id => {
-        if (deleteResumeCopy(id)) {
-          deletedCount++;
-        }
-      });
-      
-      setSavedResumes(getSavedResumes());
-      setSelectedResumeIds(new Set());
-      
-      if (deletedCount > 0) {
-        toast.success(`Deleted ${deletedCount} resume${deletedCount > 1 ? 's' : ''}`);
-      }
-    }
-  };
-
-  const toggleResumeSelection = (resumeId: string) => {
-    const newSelection = new Set(selectedResumeIds);
-    if (newSelection.has(resumeId)) {
-      newSelection.delete(resumeId);
-    } else {
-      newSelection.add(resumeId);
-    }
-    setSelectedResumeIds(newSelection);
-  };
-
-  const toggleSelectAll = () => {
-    if (selectedResumeIds.size === savedResumes.length) {
-      setSelectedResumeIds(new Set());
-    } else {
-      setSelectedResumeIds(new Set(savedResumes.map(r => r.id)));
     }
   };
 
@@ -254,97 +67,7 @@ function AppSidebar({
         <SidebarGroup>
           <SidebarGroupLabel className="text-gray-700 dark:text-zinc-300">Resume</SidebarGroupLabel>
           <SidebarGroupContent className="px-2">
-            {isEditingName ? (
-              <div className="flex items-center space-x-1">
-                <input
-                  ref={inputRef}
-                  type="text"
-                  value={tempName}
-                  onChange={(e) => setTempName(e.target.value)}
-                  onKeyDown={handleKeyDown}
-                  className="flex-1 px-2 py-1 text-sm bg-white dark:bg-zinc-700 border border-gray-300 dark:border-zinc-600 rounded text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-2 focus:ring-purple-500"
-                  placeholder="Resume name"
-                />
-                <button
-                  onClick={handleSaveEdit}
-                  className="p-1 text-green-600 hover:text-green-700 dark:text-green-400 dark:hover:text-green-300"
-                  title="Save name"
-                >
-                  <Check className="size-4" />
-                </button>
-                <button
-                  onClick={handleCancelEdit}
-                  className="p-1 text-red-600 hover:text-red-700 dark:text-red-400 dark:hover:text-red-300"
-                  title="Cancel edit"
-                >
-                  <X className="size-4" />
-                </button>
-              </div>
-            ) : (
-              <div className="flex items-center space-x-1 group">
-                <span className="flex-1 text-sm text-gray-700 dark:text-gray-300 font-bold truncate">
-                  {resumeName}
-                </span>
-                <button
-                  onClick={handleStartEdit}
-                  className="p-1 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200"
-                  title="Edit resume name"
-                >
-                  <Edit2 className="size-4" />
-                </button>
-                <DropdownMenu>
-                  <DropdownMenuTrigger asChild>
-                    <button
-                      className="p-1 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200"
-                      title="Resume options"
-                    >
-                      <MoreVertical className="size-4" />
-                    </button>
-                  </DropdownMenuTrigger>
-                  <DropdownMenuContent align="end" className="w-64">
-                    <DropdownMenuLabel className="mb-0.75 text-xs">Resume</DropdownMenuLabel>
-                    <DropdownMenuItem onClick={handleNewResume}>
-                      <File className="mr-0.5 h-4 w-4" />
-                      New
-                    </DropdownMenuItem>
-                    <DropdownMenuItem onClick={() => setIsOpenDialogOpen(true)}>
-                      <FolderOpen className="mr-0.5 h-4 w-4" />
-                      Open
-                    </DropdownMenuItem>
-                    <DropdownMenuItem onClick={handleSaveAsCopy}>
-                      <SaveAllIcon className="mr-0.5 h-4 w-4" />
-                      Save As Copy
-                    </DropdownMenuItem>
-                    <DropdownMenuSeparator />
-                    <DropdownMenuLabel className="mb-0.75 text-xs">Recent</DropdownMenuLabel>
-                    {savedResumes.length > 0 ? (
-                      <>
-                        {savedResumes.slice(0, 5).map((resume) => (
-                          <div key={resume.id} className="group/item">
-                            <DropdownMenuItem
-                              onClick={() => handleLoadResume(resume.id)}
-                              className="pr-8"
-                            >
-                              <FileText className="mr-0.5" />
-                              <div className="flex-1 overflow-hidden">
-                                <div className="truncate text-sm">{resume.name}</div>
-                                <div className="text-xs text-muted-foreground">
-                                  {formatRelativeTime(resume.lastUpdated)}
-                                </div>
-                              </div>
-                            </DropdownMenuItem>
-                          </div>
-                        ))}
-                      </>
-                    ) : (
-                      <DropdownMenuItem disabled>
-                        <span className="text-muted-foreground">No recent resumes</span>
-                      </DropdownMenuItem>
-                    )}
-                  </DropdownMenuContent>
-                </DropdownMenu>
-              </div>
-            )}
+            <ResumeManager onNewResume={handleNewResume} />
           </SidebarGroupContent>
         </SidebarGroup>
         <SidebarGroup className="block lg:hidden py-0">
@@ -453,89 +176,6 @@ function AppSidebar({
             </div>
           </SidebarGroupContent>
         </SidebarGroup>
-        
-        {/* Open Resume Dialog */}
-        <Dialog open={isOpenDialogOpen} onOpenChange={setIsOpenDialogOpen}>
-          <DialogContent className="max-w-2xl max-h-[80vh]">
-            <DialogHeader>
-              <DialogTitle>
-                <FolderOpen className="pb-1 inline me-1 size-6" strokeWidth={1.334} />
-                Open Resume
-              </DialogTitle>
-              <DialogDescription>
-                Select a saved resume to open or manage your saved copies.
-              </DialogDescription>
-            </DialogHeader>
-            <div className="space-y-4">
-              {savedResumes.length > 0 && (
-                <div className="flex items-center justify-between px-1">
-                  <div className="flex items-center space-x-2">
-                    <Checkbox
-                      checked={selectedResumeIds.size === savedResumes.length && savedResumes.length > 0}
-                      onCheckedChange={() => toggleSelectAll()}
-                    />
-                    <label className="text-sm font-medium">
-                      Select all ({savedResumes.length})
-                    </label>
-                  </div>
-                  {selectedResumeIds.size > 0 && (
-                    <Button
-                      theme="danger"
-                      text={`Delete (${selectedResumeIds.size})`}
-                      onClick={handleDeleteSelected}
-                      className="text-sm py-1.5"
-                    />
-                  )}
-                </div>
-              )}
-              <div className="overflow-y-auto max-h-[50vh]">
-                {savedResumes.length > 0 ? (
-                  <div className="space-y-2">
-                    {savedResumes.map((resume) => (
-                      <div
-                        key={resume.id}
-                        className={`group flex items-center justify-between p-3 border rounded-lg transition-colors ${
-                          selectedResumeIds.has(resume.id) 
-                            ? 'bg-blue-50 dark:bg-blue-950/30 border-blue-300 dark:border-blue-700' 
-                            : 'hover:bg-gray-50 dark:hover:bg-zinc-800'
-                        }`}
-                      >
-                        <div className="flex items-center space-x-3 flex-1 min-w-0">
-                          <Checkbox
-                            checked={selectedResumeIds.has(resume.id)}
-                            onCheckedChange={() => toggleResumeSelection(resume.id)}
-                            onClick={(e) => e.stopPropagation()}
-                          />
-                          <div 
-                            className="flex items-center space-x-3 flex-1 min-w-0 cursor-pointer"
-                            onClick={() => handleLoadResume(resume.id)}
-                          >
-                            <FileText className="h-5 w-5 text-gray-500 dark:text-gray-400 flex-shrink-0" />
-                            <div className="flex-1 min-w-0">
-                              <p className="text-sm font-medium text-gray-900 dark:text-gray-100 truncate">
-                                {resume.name}
-                              </p>
-                              <p className="text-sm text-gray-500 dark:text-gray-400">
-                                Created: {new Date(resume.createdAt).toLocaleDateString()} • 
-                                Modified: {formatRelativeTime(resume.lastUpdated)}
-                              </p>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                ) : (
-                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                    <FileText className="h-12 w-12 mx-auto mb-3 opacity-50" />
-                    <p>No saved resumes found</p>
-                    <p className="text-sm mt-1">Save your current resume using "Save As Copy" to get started</p>
-                  </div>
-                )}
-              </div>
-            </div>
-          </DialogContent>
-        </Dialog>
       </SidebarContent>
       <SidebarFooter className="text-center bg-gray-100/10 dark:bg-zinc-800/60">
         <div className="text-xs text-gray-600 dark:text-gray-400">
